requirements.txt
django
passporteye
djangorestframework